% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egEdgeroi.R
\name{CubistIAK}
\alias{CubistIAK}
\title{Run Iak3d project with building Cubist model}
\usage{
CubistIAK(datafeed)
}
\value{
an object with lmm.fit.selected,xkVal and vkVal
}
\description{
This function builds cubist model
}
\examples{
Cubistdata <- CubistIAK(datafeed)
}
