% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egEdgeroi.R
\docType{data}
\name{EdgeroiFitData}
\alias{EdgeroiFitData}
\title{Fit data example for Edgeroi Area. Expected to be DataFrame with column names:
x : x dimension (in Km)
y : y dimension (in Km)
lowerDI : lower Depth Interval of soil range in metres
upperDI : upper Depth Interval of soil range in metres
z:  response variable for valid set
dIMidPts :  Mid point of Depth Interval
Other columns relate to covariates and column names should reflect spatialCovs argument}
\description{
Fit data example for Edgeroi Area. Expected to be DataFrame with column names:
x : x dimension (in Km)
y : y dimension (in Km)
lowerDI : lower Depth Interval of soil range in metres
upperDI : upper Depth Interval of soil range in metres
z:  response variable for valid set
dIMidPts :  Mid point of Depth Interval
Other columns relate to covariates and column names should reflect spatialCovs argument
}
\keyword{data}
