% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optifix.R
\name{optifix}
\alias{optifix}
\title{optifix. Optimise with fixed parameters}
\usage{
optifix(
  par,
  fixed,
  fn,
  gr = NULL,
  ...,
  method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN"),
  lower = -Inf,
  upper = Inf,
  control = list(),
  hessian = FALSE
)
}
\description{
its like optim, but with fixed parameters.
}
\details{
specify a second argument 'fixed', a vector of TRUE/FALSE values. 
If TRUE, the corresponding parameter in fn() is fixed. Otherwise its 
variable and optimised over. 

The return thing is the return thing from optim() but with a couple of extra 
bits - a vector of all the parameters and a vector copy of the 'fixed' argument. 

Written by Barry Rowlingson  October 2011 

This file released under a CC By-SA license: 
http://creativecommons.org/licenses/by-sa/3.0/ 

and must retain the text: "Originally written by Barry Rowlingson" in comments. 

eg 
tmp <- optifix(c(1.00026,0.3),c(TRUE, FALSE), fr,method="Nelder-Mead")
}
