% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egEdgeroi.R
\name{RunPlots}
\alias{RunPlots}
\title{Run Iak3d project with building spline model}
\usage{
RunPlots(lmm.fit.selected)
}
\arguments{
\item{lmm.fit.selected}{The result of running either Cubist or Spline Models}
}
\value{
saves plots in the working directory
}
\description{
RunPlots(iakdata$lmm.fit.selected)
}
