% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egEdgeroi.R
\docType{data}
\name{Uniform_Data_Edgeroi}
\alias{Uniform_Data_Edgeroi}
\title{Represents the underlying data and required format needed to run IAK.
Example data has been included in the package, sourced from the Edgeroi Package (i.e data(edgeroi) and data(edgeroiCovariates), 
named Iak3dSIH::Uniform_Data_Edgeroi. Any data can be inputted provided the structure is as follows:}
\description{
Model Fit set includes:
cFit : Coordinates (in Km) of calib set, in two dimensions x y. 
dIFit : Depth intervals of fit set (metres) in two dimensions representing upper and lower soil sample range.
covsFit : covariates of fit set, with column names that represent covariates aligned with param spatialCovs. Within this, dIMidPts is Required (represents the mid depth i.e. average between the upper and lower depth of an observation.)
zFit : response variable of fit set
profIDFit : profile ID of fit set
Validation Set - Required if validation flag is TRUE, includes:
cVal :  Coordinates (in Km) of valid set
dIVal : depth intervals of valid set (in metres)
covsVal :  Covariates for valid set, with column names that represent covariates aligned with param spatialCovs, Within this, dIMidPts Required (represents the mid depth i.e. average between the upper and lower depth of an observation.)
zVal :  response variable for valid set
profIDVal :  profile ID representing unique strings or numbers
rList :  raster list of covariate - OPTIONAL
}
\keyword{data}
