% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egEdgeroi.R
\name{CubistIAK}
\alias{CubistIAK}
\title{Run Iak3d project with building Cubist model}
\usage{
CubistIAK(data, validation, spatialCovs, rand6ForPlot)
}
\arguments{
\item{data}{data to feed in such as Uniform_Data_Edgeroi}

\item{validation}{Boolean TRUE or FALSE}

\item{spatialCovs}{lsit of spatial covariates example c('elevation' , 'twi' , 'radK' , 'landsat_b3' , 'landsat_b4')}

\item{rand6ForPlot}{parameters passed onto the plot XX if if validation is set to TRUE}
}
\value{
an object with lmm.fit.selected,xkVal and vkVal
}
\description{
This function builds cubist model
}
\examples{
Cubistdata <- CubistIAK(Uniform_Data_Edgeroi,TRUE,c('elevation' , 'twi' , 'radK' , 'landsat_b3' , 'landsat_b4'),c(6 , 19 , 49 , 41 , 3 , 24))
}
