% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egEdgeroi.R
\name{CubistIAK}
\alias{CubistIAK}
\title{Run Iak3d project with building Cubist model}
\usage{
CubistIAK(data = data, validation = validation, spatialCovs = spatialCovs)
}
\arguments{
\item{data}{data to feed in such as Uniform_Data_Edgeroi. Please see associated documentation for expected structure.}

\item{validation}{Boolean TRUE or FALSE to run validation. If True, columns relating to validation is needed in the data. i.e. cVal, DIVal, covsVal etc.}

\item{spatialCovs}{lsit of spatial covariates example c('elevation' , 'twi' , 'radK' , 'landsat_b3' , 'landsat_b4')}
}
\value{
an object with lmm.fit.selected,xkVal and vkVal
}
\description{
This function builds cubist model
}
\examples{
Cubistdata <- CubistIAK(Uniform_Data_Edgeroi,TRUE,c('elevation' , 'twi' , 'radK' , 'landsat_b3' , 'landsat_b4'))
}
