% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egEdgeroi.R
\name{RunPlots}
\alias{RunPlots}
\title{Run Iak3d project with building spline model. Use after running either SplineIAK OR CubistIAK functions.
Example: RunPlots(fit = Cubistdata$lmm.fit.selected, InputParamatersList = Cubistdata$InputParamatersList,c(1,2,3,4,5,6))}
\usage{
RunPlots(
  fit = lmm.fit.selected,
  InputParamatersList = InputParamatersList,
  chooseToPlot = chooseToPlot
)
}
\arguments{
\item{InputParamatersList}{Included in the result of running either Cubist or Spline Models. Attaches as input run to an output result and enables plots on the outputs for analysis.}

\item{chooseToPlot}{Select a maximum of six co-ordinate points to plot for inspection, example: c(1,2,3,4,5,6). Plot is calebrated for 6 plots. These are rrediction plots that are saved in plotVal4Plot.pdf. Upper limit Limited by unique number of cVal.}

\item{lmm.fit.selected}{Included in the result of running either Cubist or Spline Mode.}
}
\value{
saves plots in the working directory
}
\description{
Run Iak3d project with building spline model. Use after running either SplineIAK OR CubistIAK functions.
Example: RunPlots(fit = Cubistdata$lmm.fit.selected, InputParamatersList = Cubistdata$InputParamatersList,c(1,2,3,4,5,6))
}
