% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egEdgeroi.R
\name{SplineIAK}
\alias{SplineIAK}
\title{Run Iak3d project with building spline model}
\usage{
SplineIAK(datafeed)
}
\value{
an object with lmm.fit.selected,xkVal and vkVal
}
\description{
This function builds spline model
}
\examples{
Splinedata <- SplineIAK(datafeed)
}
