% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egEdgeroi.R
\name{SplineIAK}
\alias{SplineIAK}
\title{This function builds spline model from a uniform data structure available in package (Uniform_Data_Edgeroi).}
\usage{
SplineIAK(datafeedin)
}
\value{
an object with lmm.fit.selected,xkVal and vkVal
}
\description{
This function builds spline model from a uniform data structure available in package (Uniform_Data_Edgeroi).
}
\examples{
Splinedata <- SplineIAK(Uniform_Data_Edgeroi)
}
