% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egEdgeroi.R
\name{SplineIAK}
\alias{SplineIAK}
\title{This function builds spline model from a uniform data structure available in package (Uniform_Data_Edgeroi).}
\usage{
SplineIAK(data = data, validation = validation, spatialCovs = spatialCovs)
}
\arguments{
\item{data}{data to feed in such as Uniform_Data_Edgeroi}

\item{validation}{Boolean TRUE or FALSE to run validation. If True, columns relating to validation is needed in the data. i.e. cVal, DIVal etc.}

\item{spatialCovs}{list of spatial covariates, i.e  c('elevation' , 'twi' , 'radK' , 'landsat_b3' , 'landsat_b4') For Edgeroi Dataset}
}
\value{
an object with lmm.fit.selected,xkVal and vkVal
}
\description{
This function builds spline model from a uniform data structure available in package (Uniform_Data_Edgeroi).
}
\examples{
Splinedata <- SplineIAK(Uniform_Data_Edgeroi,TRUE,c('elevation' , 'twi' , 'radK' , 'landsat_b3' , 'landsat_b4'))
}
