% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egEdgeroi.R
\name{SplineIAK}
\alias{SplineIAK}
\title{This function builds spline model from a uniform data structure available in package (Uniform_Data_Edgeroi).}
\usage{
SplineIAK(
  fit_data = fit_data,
  validation = validation,
  validate_data = validate_data,
  spatialCovs = spatialCovs
)
}
\arguments{
\item{fit_data}{Fit data to feed in such as EdgeroiFitData. Expected to be a dataframe.}

\item{validation}{Boolean TRUE or FALSE to run validation. If True, columns relating to validation is needed in the data. i.e. cVal, DIVal etc.}

\item{validate_data}{Validation data to feed in such as EdgeroiValidationData. Expected to be a dataframe.}

\item{spatialCovs}{list of spatial covariates found in covariates data i.e covsFit and covsVal. Can be any subset of these covariates. For example  c('elevation' , 'twi' , 'radK' , 'landsat_b3' , 'landsat_b4') For Edgeroi Dataset}
}
\value{
an object with lmm.fit.selected,xkVal and vkVal
}
\description{
This function builds spline model from a uniform data structure available in package (Uniform_Data_Edgeroi).
}
\examples{
Splinedata <- SplineIAK(fit_data = EdgeroiFitData, validation = TRUE,validate_data = EdgeroiValidationData, spatialCovs = c('elevation' , 'twi' , 'radK' , 'landsat_b3' , 'landsat_b4'))
}
