% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egEdgeroi.R
\name{SplineIAK}
\alias{SplineIAK}
\title{This function builds spline model from a uniform data structure available in package (Uniform_Data_Edgeroi).}
\usage{
SplineIAK(
  fit_data = fit_data,
  validate_data = validate_data,
  spatialCovs = spatialCovs
)
}
\arguments{
\item{fit_data}{Fit data to feed in such as EdgeroiFitData. Expected to be a dataframe.}

\item{validate_data}{OPTIONAL Validation data to feed in such as EdgeroiValidationData. Expected to be a dataframe.}

\item{spatialCovs}{list of spatial covariates with 'dIMidPts' required. example c(''dIMidPts',elevation' , 'twi' , 'radK' , 'landsat_b3' , 'landsat_b4').}
}
\value{
an object with lmm.fit.selected,xkVal and vkVal
}
\description{
This function builds spline model from a uniform data structure available in package (Uniform_Data_Edgeroi).
}
\examples{
Splinedata <- SplineIAK(fit_data = EdgeroiFitData,validate_data = EdgeroiValidationData, spatialCovs = c('dIMidPts','elevation' , 'twi' , 'radK' , 'landsat_b3' , 'landsat_b4'))
}
