% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egEdgeroi.R
\docType{data}
\name{Uniform_Data_Edgeroi}
\alias{Uniform_Data_Edgeroi}
\title{Represents the underlying data needed to run Spline or Cubist Models within this package.
Example data has been included in the package, sourced from the Edgeroi Package (i.e data(edgeroi) and data(edgeroiCovariates), 
named Iak3dSIH::Uniform_Data_Edgeroi. Any data can be inputted provided the structure is as follows:}
\description{
Represents the underlying data needed to run Spline or Cubist Models within this package.
Example data has been included in the package, sourced from the Edgeroi Package (i.e data(edgeroi) and data(edgeroiCovariates), 
named Iak3dSIH::Uniform_Data_Edgeroi. Any data can be inputted provided the structure is as follows:
}
\keyword{data}
